+++ src/helpers/FrameworkDetector.ts
@@ -1,5 +1,5 @@
-import { existsSync } from "fs";
-import { resolve } from "path";
+import { existsSync, readFileSync } from "fs";
+import { join, resolve } from "path";
 import { FrameworkDetectors } from "../constants/FrameworkDetectors";
 import { Extension } from "./Extension";
 
@@ -14,16 +14,50 @@ export class FrameworkDetector {
   }
 
   private static check(folder: string) {
-    const { dependencies, devDependencies } = Extension.getInstance().packageJson;
+    let dependencies = null;
+    let devDependencies = null;
+    let gemContent = null;
+
+    // Try fetching the package JSON file
+    try {
+      const pkgFile = join(folder, 'package.json');
+      if (existsSync(pkgFile)) {
+        let packageJson: any = readFileSync(pkgFile, "utf8");
+        if (packageJson) {
+          packageJson = typeof packageJson === "string" ? JSON.parse(packageJson) : packageJson;
+
+          dependencies = packageJson.dependencies || null;
+          devDependencies = packageJson.devDependencies || null;
+        }
+      }
+    } catch (e) {
+      // do nothing
+    }
+
+    // Try fetching the Gemfile
+    try {
+      const gemFile = join(folder, 'Gemfile');
+      if (existsSync(gemFile)) {
+        gemContent = readFileSync(gemFile, "utf8");
+      }
+    } catch (e) {
+      // do nothing
+    }
 
     for (const detector of FrameworkDetectors) {
       if (detector && folder) {
 
         // Verify by dependencies
         for (const dependency of detector.requiredDependencies ?? []) {
+          // Checks for package.json dependencies
           const inDependencies = dependencies && dependencies[dependency]
           const inDevDependencies = devDependencies && devDependencies[dependency]
           if (inDependencies || inDevDependencies) {
+            return detector.framework;
+          }
+
+          // Checks for Gemfile
+          if (gemContent && gemContent.includes(dependency)) {
             return detector.framework;
           }
         }